Fix various bugs

+++ MessageHider/BasicSteganoR.cpp
@@ -12,7 +12,7 @@ BasicSteganoR::BasicSteganoR()
 
 int BasicSteganoR::MessageMaxLenght(const size_t len_pixel_tab, int nb_bytes_per_pixel)
 {
-	return len_pixel_tab / 8;
+	return static_cast<int>(len_pixel_tab / 8);
 }
 
 void BasicSteganoR::Encrypt(unsigned char* pixel_tab, const size_t len_pixel_tab, int nb_bytes_per_pixel, std::string word_to_hide)
